Added some comments & put RunTime in seperate file
